chore: move handling of if_exists to base.py
